Add files via upload
Simple quiz that runs True or False Trivia questions from any DB that manage to export the data through Json files(due to similarity between Json datasets and python list/dictionaries) you have to just make a few tweaks to the names of the attributes of the code or data to make sure that the methods are able to pick-up the data. Still, it has no GUI and runs on the terminal. Entirely coded using OOP and very easy to understand if you're familiar with these types of stuff.